Fix travis build

+++ src/FontInfo.cpp
@@ -63,7 +63,7 @@ void FontInfo::writeToXmlFile(const std::string &fileName) const
     for (size_t i = 0; i < pages.size(); ++i)
     {
         tinyxml2::XMLElement* pageElement = doc.NewElement("page");
-        pageElement->SetAttribute("id", i);
+        pageElement->SetAttribute("id", static_cast<unsigned int>(i));
         pageElement->SetAttribute("file", pages[i].c_str());
         pagesElement->InsertEndChild(pageElement);
     }
